% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulkRNAseq_ABIS}
\alias{bulkRNAseq_ABIS}
\title{PBMCs expression profiles (ABIS dataset)}
\format{
A matrix with 1296 rows (genes) and 12 variables (samples)
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse107011}{
GEO}
}
\usage{
data(bulkRNAseq_ABIS)
}
\description{
Public dataset (GSE107011) containing the TPM-normalized gene
expression values from bulk RNAseq of PBMCs of 12 healthy individuals. We 
include here only genes selected in the signature matrices.
}
\references{
Monaco et al. (2019) Cell Reports 26, 1627–1640
(\href{https://www.cell.com/cell-reports/pdf/S2211-1247(19)30059-2.pdf}{Cell
Reports})
}
\keyword{datasets}
