% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decon_methods.R
\name{get_decon_methods}
\alias{get_decon_methods}
\title{Deconvolution methods acronyms}
\usage{
get_decon_methods()
}
\value{
vector containing the acronyms of deconvolution methods.
}
\description{
\code{get_decon_methods} returns supported deconvolution 
methods acronyms.
}
\examples{
# get available deconvolution methods
get_decon_methods()

}
\author{
Vincent Kuettel, Sabina Pfister
}
