% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundTruth_ABIS}
\alias{groundTruth_ABIS}
\title{PBMCS true cell type proprotions (ABIS dataset)}
\format{
A matrix with 12 rows (samples) and 24 variables (cell types)
}
\source{
\href{https://github.com/giannimonaco/ABIS/tree/master/data}{Github}
}
\usage{
data(groundTruth_ABIS)
}
\description{
Public dataset (GSE107011)containing the true proportions for 
all combinations of cell types (PBMCs) for 12 individuals.
}
\references{
Monaco et al. (2019) Cell Reports 26, 1627–1640
(\href{https://www.cell.com/cell-reports/pdf/S2211-1247(19)30059-2.pdf}{Cell 
Reports})
}
\keyword{datasets}
