% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ABIS.R
\name{load_ABIS}
\alias{load_ABIS}
\title{Load demo PBMCs deconvolution data}
\usage{
load_ABIS()
}
\value{
Returns string confirming successfull loading of the data.
}
\description{
\code{load_ABIS} is used to load a demo dataset for the
deconvolution of PBMCs samples from published data under the accession
number GSE107011. The dataset consists of the following datasets: 
\itemize{
\item{bulkRNAseq_ABIS: PBMCs expression profiles}
\item{sigMatrix_ABIS_S0: Signature matrix for deconvolution of PBMCs 
in 17 cell types}
\item{sigMatrix_ABIS_S1: Signature matrix for deconvolution of PBMCs
in 13 cell types}
\item{sigMatrix_ABIS_S2: Signature matrix for deconvolution of PBMCs
in 11 cell types}
\item{sigMatrix_ABIS_S3: Signature matrix for deconvolution of PBMCs
in 9 cell types}
\item{groundTruth_ABIS: PBMCS true cell type proprotions}
}
}
\examples{
# load data
load_ABIS()

}
\author{
Vincent Kuettel, Sabina Pfister
}
