% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigMatrix_ABIS_S3}
\alias{sigMatrix_ABIS_S3}
\title{Signature matrix for deconvolution of PBMCs in 9 cell types}
\format{
A matrix with 1296 rows (genes) and 9 variables (cell types)
}
\usage{
data(sigMatrix_ABIS_S3)
}
\description{
A dataset containing the TPM-normalized RNA-seq gene expression
values for signature genes of 17 PBMCs.
}
\references{
Monaco et al. (2019) Cell Reports 26, 1627–1640
(\href{https://www.cell.com/cell-reports/pdf/S2211-1247(19)30059-2.pdf}{Cell 
Reports})
}
\keyword{datasets}
